Added VCO Study
This is the current simulation model for the VCO, it has two designs: One design is for low frequency, the other design is for higher (all the other) audio frequencies.  You can directly run the simulation as is, varying the VP parametre from 0 to 5 volts will produce a varying oscillator.  You can uncomment the ".step param VP list .1 5" to see the full range of both oscillators.