Added VCA Study
This is the Voltage Controlled Amplifier, it is used to attenuate the final output signal before the power amplifier stage.  It is usually driven by the envelope generator which is gated by the keys or MIDI signal.  This module still requires some tweaking but is working for 2.5Vrms input signals, any high and it will distort at full gain, but that could be desireable as well.